Capstone_project/Invoice-Parser/requirements.txt
faiss-cpu==1.13.2
sentence-transformers
transformers

streamlit==1.32.2
pytesseract==0.3.13
Pillow==9.5.0
opencv-python==4.9.0.80
pdf2image==1.17.0

torch==2.2.2
transformers==4.38.2
sentencepiece==0.1.99

numpy==1.26.4
pandas==2.2.1
